Added project config header
To be included in "main.cpp" to provide project-wide confuration information to be used within the project itself.

This file is auto-generated by the project configuation (it is dafe to delete it entirely before running the configuration step). To add or edit the contents of this file, please specify in the associated "...Config.h.in" file, and re-run the project configuration step.